requirements.txt
streamlit==1.12.1
qrcode[pil]==6.1
Pillow==7.2.0
altair==4.1.0
